Refactor and break stuff
