+++ configs/2-be-updated/blood_rage.json
