+++ configs/2-be-updated/eldritch_horror.json
